tidy the code of db-common and db-worker
