+++ light-client-db-common/src/lib.rs
@@ -1,12 +1,15 @@
+use anyhow::anyhow;
 use anyhow::Context;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use web_sys::js_sys::{Atomics, Int32Array, Uint8Array};
 #[derive(Serialize, Deserialize, Debug)]
+/// Represent a key-value pair
 pub struct KV {
     pub key: Vec<u8>,
     pub value: Vec<u8>,
 }
 #[derive(Serialize, Deserialize, Default, Debug)]
+/// A serializable CursorDirection
 pub enum CursorDirection {
     #[default]
     Next,
@@ -15,6 +18,7 @@ pub enum CursorDirection {
     PrevUnique,
 }
 #[derive(Serialize, Deserialize, Debug)]
+/// Response of DbCommandRequest. For details, please refer to the doc of DbCommandRequest
 pub enum DbCommandResponse {
     Read { values: Vec<Option<Vec<u8>>> },
     Put,
@@ -25,22 +29,35 @@ pub enum DbCommandResponse {
 }
 
 #[derive(Deserialize, Serialize, Debug)]
+/// Represent a database command
 pub enum DbCommandRequest {
+    /// Read the value corresponding to a series of keys
+    /// Input: A series of keys
+    /// Output: A series of values corresponding to keys, None if the key wasn't found in database
     Read {
         keys: Vec<Vec<u8>>,
     },
+    /// Write a series of key-value pairs into database
+    /// Input: A series of key-value pairs
+    /// Output: None
     Put {
         kvs: Vec<KV>,
     },
+    /// Remove a series of entries from database
+    /// Input: Keys to remove
+    /// Output: None
     Delete {
         keys: Vec<Vec<u8>>,
     },
+    /// Gets at most `limit` entries, starting from `start_key_bound`, skipping the first `skip` entries, keep fetching until `take_while` evals to false
+    /// Output: Key value pairs fetched
     Iterator {
         start_key_bound: Vec<u8>,
         order: CursorDirection,
         limit: usize,
         skip: usize,
     },
+    /// Similar to `Iterator`, but only keys are returned
     IteratorKey {
         start_key_bound: Vec<u8>,
         order: CursorDirection,
@@ -52,11 +69,23 @@ pub enum DbCommandRequest {
     },
 }
 #[repr(i32)]
+/// Represent a 4-byte command which will be put in input buffer
 pub enum InputCommand {
+    /// Indicates that there is no command and light client is waiting for next call
+    /// Payload: None
     Waiting = 0,
+    /// Open database,
+    /// Payload: database name (string), bincode encoded
     OpenDatabase = 1,
+    /// Execute a database command
+    /// Payload: DbCommandRequest, bincode encoded
     DbRequest = 2,
+    /// Shutdown db worker, exiting main loop
+    /// Payload: None
+    /// Note: There won't be a response for [`crate::InputCommand::Shutdown`]
     Shutdown = 3,
+    /// Used for response from take_while, not for users
+    /// Payload: result of the call, in bool, bincode-encoded
     ResponseTakeWhile = 20,
 }
 
@@ -76,12 +105,22 @@ impl TryFrom<i32> for InputCommand {
 }
 
 #[repr(i32)]
+/// Represent a 4-byte command which will be put in output buffer
 pub enum OutputCommand {
+    /// Waiting for db worker to handle the command
+    /// Payload: None
     Waiting = 0,
+    /// Successful response of OpenDatabase
+    /// Payload: None
     OpenDatabaseResponse = 1,
+    /// Successful response of DbRequest
+    /// Payload: bincode-encoded DbCommandResponse
     DbResponse = 2,
-    ShutdownResponse = 3,
+    /// Error of OpenDatabaseRequest or DbRequest
+    /// Payload: bincode-encoded string
     Error = 10,
+    /// DbWorker wants to call take_while
+    /// Payload: bincode-encoded bytes, argument of take_while
     RequestTakeWhile = 20,
 }
 
@@ -93,14 +132,13 @@ impl TryFrom<i32> for OutputCommand {
             0 => Ok(Self::Waiting),
             1 => Ok(Self::OpenDatabaseResponse),
             2 => Ok(Self::DbResponse),
-            3 => Ok(Self::ShutdownResponse),
             10 => Ok(Self::Error),
             20 => Ok(Self::RequestTakeWhile),
             s => Err(anyhow!("Invalid command: {}", s)),
         }
     }
 }
-
+/// Translate a [`crate::CursorDirection`] to [`idb::CursorDirection`]
 pub fn ckb_cursor_direction_to_idb(x: crate::CursorDirection) -> idb::CursorDirection {
     use crate::CursorDirection;
     match x {
@@ -110,6 +148,7 @@ pub fn ckb_cursor_direction_to_idb(x: crate::CursorDirection) -> idb::CursorDire
         CursorDirection::PrevUnique => idb::CursorDirection::PrevUnique,
     }
 }
+/// Translate a [`idb::CursorDirection`] to [`crate::CursorDirection`]
 pub fn idb_cursor_direction_to_ckb(x: idb::CursorDirection) -> crate::CursorDirection {
     use crate::CursorDirection;
 
@@ -120,8 +159,13 @@ pub fn idb_cursor_direction_to_ckb(x: idb::CursorDirection) -> crate::CursorDire
         idb::CursorDirection::PrevUnique => CursorDirection::PrevUnique,
     }
 }
-use anyhow::anyhow;
-
+/// Fill a input buffer/output buffer with a [`crate::InputCommand`]/[`crate::OutputCommand`] and the buffer
+/// The buffer would be in 4byte command + 4byte payload length + payload
+///
+/// cmd: Command in i32
+/// data: The payload of command
+/// i32arr: Int32Array view of the buffer
+/// u8arr: Uint8Array view of the buffer
 pub fn write_command_with_payload<T: Serialize>(
     cmd: i32,
     data: T,
@@ -139,6 +183,9 @@ pub fn write_command_with_payload<T: Serialize>(
     Atomics::notify(i32arr, 0).map_err(|e| anyhow!("Failed to notify: {e:?}"))?;
     Ok(())
 }
+/// Read the payload from the given input buffer/output buffer
+/// i32arr: Int32Array view of the buffer
+/// u8arr: Uint8Array view of the buffer
 pub fn read_command_payload<T: DeserializeOwned>(
     i32arr: &Int32Array,
     u8arr: &Uint8Array,
